dividend-history/requirements.txt
aiodns==2.0.0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
bs4==0.0.1
ccxt==1.42.18
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
Flask==1.1.2
graphviz==0.13.2
gunicorn==20.0.4
idna==2.10
install==1.3.4
isort==5.7.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
metatrader==0.0.1
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.1
pandas==1.2.2
parse==1.19.0
Pillow==8.1.0
polygon-api-client==0.1.9
print==1.3.0
protobuf==3.14.0
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
pyee==8.1.0
pylint==2.7.1
pyOMT5==0.0.29
pyotp==2.4.1
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
requests-html==0.10.0
robin-stocks==1.6.2
scipy==1.6.1
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
ta==0.7.0
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
w3lib==1.22.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yahoo-earnings-calendar==0.6.0
yahoo-fin==0.8.7
yarl==1.1.0
yfinance==0.1.55
